requirements.txt
opencv-python
numpy
Pillow
imagehash
tqdm
pandas
scikit-image
scikit-learn
matplotlib
openpyxl
